feat: Enhance AppendCommand with context and meta improvements
AppendCommand now respects the current context and can take default meta data.
Commands set the base meta data to include the id of the frame which registered
the command and the id of the frame which triggered the command call. Added a
test to assert call .append from a command works correctly.